backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
sqlalchemy>=2.0.30
oracledb>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.2.1
slowapi>=0.1.9
python-multipart>=0.0.9
python-json-logger>=2.0.7
prometheus-fastapi-instrumentator>=6.1.0
psutil>=5.9.0
